Cypress custom commands (#365)
* [FEATURE] Detector must have at least one alert set #288


* [FEATURE] Add cypress custom commands to support better type events #364


* [FEATURE] Add cypress custom commands to support better type events #364


* [FEATURE] Add cypress custom commands to support better type events #364


* [FEATURE] Add cypress custom commands to support better type events #364


* [FEATURE] Add cypress custom commands to support better type events #364


* [FEATURE] Add cypress custom commands to support better type events #364


* [FEATURE] Add cypress custom commands to support better type events #364


* [FEATURE] Add cypress custom commands to support better type events #364
