+++ cypress/integration/1_detectors.spec.js
@@ -72,8 +72,10 @@ describe('Detectors', () => {
       .realType(indexName);
 
     cy.intercept({
-      method: 'POST',
       pathname: '/_plugins/_security_analytics/rules/_search',
+      query: {
+        prePackaged: 'true',
+      },
     }).as('getSigmaRules');
 
     // Select threat detector type (Windows logs)
@@ -84,7 +86,7 @@ describe('Detectors', () => {
       cy.get('[data-test-subj="detection-rules-btn"]').click({ force: true, timeout: 5000 });
 
       // find search, type USB
-      cy.triggerSearchField('Search...', 'USB Device Plugged');
+      cy.get(`input[placeholder="Search..."]`).ospSearch('USB Device Plugged');
 
       // Disable all rules
       cy.contains('tr', 'USB Device Plugged', { timeout: 1000 });
@@ -207,8 +209,8 @@ describe('Detectors', () => {
 
     // Change detector name
     cy.get(`input[placeholder="Enter a name for the detector."]`)
-      .clearInput()
-      .focus()
+      .realClick()
+      .ospClear()
       .realType('test detector edited');
 
     // Change detector description
@@ -219,16 +221,13 @@ describe('Detectors', () => {
     // Change input source
     cy.get(`[data-test-subj="define-detector-select-data-source"]`)
       .find('input')
-      .clearInput()
+      .ospClear()
       .focus()
       .realType('.opensearch-notifications-config')
       .realPress('Enter');
 
     // Change detector scheduling
-    cy.get(`[data-test-subj="detector-schedule-number-select"]`)
-      .clearInput()
-      .focus()
-      .realType('10');
+    cy.get(`[data-test-subj="detector-schedule-number-select"]`).ospClear().focus().realType('10');
     cy.get(`[data-test-subj="detector-schedule-unit-select"]`).select('Hours');
 
     // Save changes to detector details
@@ -270,7 +269,7 @@ describe('Detectors', () => {
     });
 
     // Search for specific rule
-    cy.triggerSearchField('Search...', 'USB Device');
+    cy.get(`input[placeholder="Search..."]`).ospSearch('USB Device');
 
     // Toggle single search result to unchecked
     cy.contains('table tr', 'USB Device Plugged').within(() => {
@@ -294,7 +293,7 @@ describe('Detectors', () => {
     });
 
     // Search for specific rule
-    cy.triggerSearchField('Search...', 'USB');
+    cy.get(`input[placeholder="Search..."]`).ospSearch('USB');
 
     // Toggle single search result to checked
     cy.contains('table tr', 'USB Device Plugged').within(() => {
