+++ cypress/integration/4_findings.spec.js
@@ -56,7 +56,7 @@ describe('Findings', () => {
 
   it('displays finding details flyout when user clicks on View details icon', () => {
     // filter table to show only sample_detector findings
-    cy.triggerSearchField('Search findings', 'sample_detector');
+    cy.get(`input[placeholder="Search findings"]`).ospSearch('sample_detector');
 
     // Click View details icon
     cy.getTableFirstRow('[data-test-subj="view-details-icon"]').then(($el) => {
@@ -73,7 +73,7 @@ describe('Findings', () => {
 
   it('displays finding details flyout when user clicks on Finding ID', () => {
     // filter table to show only sample_detector findings
-    cy.triggerSearchField('Search findings', 'sample_detector');
+    cy.get(`input[placeholder="Search findings"]`).ospSearch('sample_detector');
 
     // Click findingId to trigger Finding details flyout
     cy.getTableFirstRow('[data-test-subj="finding-details-flyout-button"]').then(($el) => {
@@ -90,7 +90,7 @@ describe('Findings', () => {
 
   it('allows user to view details about rules that were triggered', () => {
     // filter table to show only sample_detector findings
-    cy.triggerSearchField('Search findings', 'sample_detector');
+    cy.get(`input[placeholder="Search findings"]`).ospSearch('sample_detector');
 
     // open Finding details flyout via finding id link. cy.wait essential, timeout insufficient.
     cy.get(`[data-test-subj="view-details-icon"]`).eq(0).click({ force: true });
@@ -116,7 +116,7 @@ describe('Findings', () => {
 
   it('opens rule details flyout when rule name inside accordion drop down is clicked', () => {
     // filter table to show only sample_detector findings
-    cy.triggerSearchField('Search findings', 'sample_detector');
+    cy.get(`input[placeholder="Search findings"]`).ospSearch('sample_detector');
 
     // open Finding details flyout via finding id link. cy.wait essential, timeout insufficient.
     cy.getTableFirstRow('[data-test-subj="view-details-icon"]').then(($el) => {
@@ -142,7 +142,7 @@ describe('Findings', () => {
     });
 
     // filter table to show only sample_detector findings
-    cy.triggerSearchField('Search threat detectors', 'sample_detector');
+    cy.get(`input[placeholder="Search threat detectors"]`).ospSearch('sample_detector');
 
     // intercept detectors and rules requests
     cy.intercept('detectors/_search').as('getDetector');
@@ -166,7 +166,7 @@ describe('Findings', () => {
           cy.get('[data-test-subj="editButton"]').contains('Delete').click({ force: true });
 
           // Search for sample_detector, presumably deleted
-          cy.triggerSearchField('Search threat detectors', 'sample_detector');
+          cy.get(`input[placeholder="Search threat detectors"]`).ospSearch('sample_detector');
 
           // Confirm sample_detector no longer exists
           cy.contains('There are no existing detectors.');
